integrations/polymarket/requirements.txt
py-clob-client>=0.15.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
python-dotenv>=1.0.0
rich>=13.7.0
apscheduler>=3.10.0
tavily-python>=0.3.0
requests>=2.31.0
websockets>=12.0
prometheus-client>=0.20.0
